+++ storage_drivers/ontap/ontap_san.go
@@ -151,6 +151,10 @@ func (d *SANStorageDriver) validate(ctx context.Context) error {
 		return fmt.Errorf("driver validation failed: %v", err)
 	}
 
+	if err := ValidateStoragePrefix(*d.Config.StoragePrefix); err != nil {
+		return err
+	}
+
 	if err := ValidateStoragePools(ctx, d.physicalPools, d.virtualPools, d.Name()); err != nil {
 		return fmt.Errorf("storage pool validation failed: %v", err)
 	}